+++ lisp/list-all.el
@@ -11,10 +11,6 @@
 ;;
 ;;    [--depth]  dependency level to print
 ;;
-;;  Effective flags:
-;;
-;;    [-g, --global]
-;;
 
 ;;; Code: