+++ lisp/load-path.el
@@ -6,11 +6,6 @@
 ;;
 ;;   $ eask load-path
 ;;
-;;
-;;  Effective flag:
-;;
-;;    [-g, --global]
-;;
 
 ;;; Code: